Stats: Moving Devices to use new structure
