+++ client/my-sites/stats/features/modules/stats-module-devices/index.ts
@@ -1,0 +1,2 @@
+export { default } from './stats-module-devices-wrapper';
+export { default as StatsModuleUpgradeDevicesOverlay } from './stats-module-upgrade-overlay';
