+++ client/my-sites/stats/features/modules/stats-module-devices/stats-module-devices.scss
