+++ client/my-sites/stats/features/modules/stats-module-devices/stats-module-devices.tsx
@@ -8,11 +8,11 @@ import PieChart from 'calypso/components/pie-chart';
 import PieChartLegend from 'calypso/components/pie-chart/legend';
 import { useSelector } from 'calypso/state';
 import { getSelectedSiteId } from 'calypso/state/ui/selectors';
-import useModuleDevicesQuery, { StatsDevicesData } from '../hooks/use-modeule-devices-query';
-import { QueryStatsParams } from '../hooks/utils';
-import StatsListCard from '../stats-list/stats-list-card';
-import StatsModulePlaceholder from '../stats-module/placeholder';
-import statsStrings from '../stats-strings';
+import useModuleDevicesQuery, { StatsDevicesData } from '../../../hooks/use-modeule-devices-query';
+import { QueryStatsParams } from '../../../hooks/utils';
+import StatsListCard from '../../../stats-list/stats-list-card';
+import StatsModulePlaceholder from '../../../stats-module/placeholder';
+import statsStrings from '../../../stats-strings';
 
 import './stats-module-devices.scss';
